+++ couchpotato/core/plugins/renamer/main.py
@@ -1,13 +1,15 @@
 from couchpotato import get_session
 from couchpotato.api import addApiView
 from couchpotato.core.event import addEvent, fireEvent, fireEventAsync
-from couchpotato.core.helpers.encoding import toUnicode
+from couchpotato.core.helpers.encoding import toUnicode, ss
 from couchpotato.core.helpers.request import jsonified
-from couchpotato.core.helpers.variable import getExt, mergeDicts, getTitle
+from couchpotato.core.helpers.variable import getExt, mergeDicts, getTitle, \
+    getImdb
 from couchpotato.core.logger import CPLog
 from couchpotato.core.plugins.base import Plugin
-from couchpotato.core.settings.model import Library, File, Profile
+from couchpotato.core.settings.model import Library, File, Profile, Release
 from couchpotato.environment import Env
+import errno
 import os
 import re
 import shutil
@@ -19,6 +21,7 @@ log = CPLog(__name__)
 class Renamer(Plugin):
 
     renaming_started = False
+    checking_snatched = False
 
     def __init__(self):
 
@@ -27,9 +30,15 @@ class Renamer(Plugin):
         })
 
         addEvent('renamer.scan', self.scan)
+        addEvent('renamer.check_snatched', self.checkSnatched)
+
         addEvent('app.load', self.scan)
 
-        fireEvent('schedule.interval', 'renamer.scan', self.scan, minutes = self.conf('run_every'))
+        if self.conf('run_every') > 0:
+            fireEvent('schedule.interval', 'renamer.check_snatched', self.checkSnatched, minutes = self.conf('run_every'))
+
+        if self.conf('force_every') > 0:
+            fireEvent('schedule.interval', 'renamer.check_snatched_forced', self.scan, hours = self.conf('force_every'))
 
     def scanView(self):
 
@@ -45,7 +54,7 @@ class Renamer(Plugin):
             return
 
         if self.renaming_started is True:
-            log.info('Renamer is disabled to avoid infinite looping of the same error.')
+            log.info('Renamer is already running, if you see this often, check the logs above for errors.')
             return
 
         # Check to see if the "to" folder is inside the "from" folder.
@@ -124,6 +133,8 @@ class Renamer(Plugin):
                      'resolution_width': group['meta_data'].get('resolution_width'),
                      'resolution_height': group['meta_data'].get('resolution_height'),
                      'imdb_id': library['identifier'],
+                     'cd': '',
+                     'cd_nr': '',
                 }
 
                 for file_type in group['files']:
@@ -141,7 +152,7 @@ class Renamer(Plugin):
                         continue
 
                     # Move other files
-                    multiple = len(group['files']['movie']) > 1 and not group['is_dvd']
+                    multiple = len(group['files'][file_type]) > 1 and not group['is_dvd']
                     cd = 1 if multiple else 0
 
                     for current_file in sorted(list(group['files'][file_type])):
@@ -154,24 +165,19 @@ class Renamer(Plugin):
                         replacements['ext'] = getExt(current_file)
 
                         # cd #
-                        replacements['cd'] = ' cd%d' % cd if cd else ''
-                        replacements['cd_nr'] = cd
+                        replacements['cd'] = ' cd%d' % cd if multiple else ''
+                        replacements['cd_nr'] = cd if multiple else ''
 
                         # Naming
                         final_folder_name = self.doReplace(folder_name, replacements)
                         final_file_name = self.doReplace(file_name, replacements)
                         replacements['filename'] = final_file_name[:-(len(getExt(final_file_name)) + 1)]
 
-                        # Group filename without cd extension
-                        replacements['cd'] = ''
-                        replacements['cd_nr'] = ''
-                        group['filename'] = self.doReplace(file_name, replacements)[:-(len(getExt(final_file_name)) + 1)]
-
                         # Meta naming
                         if file_type is 'trailer':
-                            final_file_name = self.doReplace(trailer_name, replacements)
+                            final_file_name = self.doReplace(trailer_name, replacements, remove_multiple = True)
                         elif file_type is 'nfo':
-                            final_file_name = self.doReplace(nfo_name, replacements)
+                            final_file_name = self.doReplace(nfo_name, replacements, remove_multiple = True)
 
                         # Seperator replace
                         if separator:
@@ -202,9 +208,15 @@ class Renamer(Plugin):
                         # Check for extra subtitle files
                         if file_type is 'subtitle':
 
-                            # rename subtitles with or without language
-                            rename_files[current_file] = os.path.join(destination, final_folder_name, final_file_name)
+                            remove_multiple = False
+                            if len(group['files']['movie']) == 1:
+                                remove_multiple = True
+
                             sub_langs = group['subtitle_language'].get(current_file, [])
+
+                            # rename subtitles with or without language
+                            sub_name = self.doReplace(file_name, replacements, remove_multiple = remove_multiple)
+                            rename_files[current_file] = os.path.join(destination, final_folder_name, sub_name)
 
                             rename_extras = self.getRenameExtras(
                                 extra_type = 'subtitle_extra',
@@ -213,20 +225,19 @@ class Renamer(Plugin):
                                 file_name = file_name,
                                 destination = destination,
                                 group = group,
-                                current_file = current_file
+                                current_file = current_file,
+                                remove_multiple = remove_multiple,
                             )
 
-                            # Don't add language if multiple languages in 1 file
-                            if len(sub_langs) > 1:
-                                rename_files[current_file] = os.path.join(destination, final_folder_name, final_file_name)
-                            elif len(sub_langs) == 1:
+                            # Don't add language if multiple languages in 1 subtitle file
+                            if len(sub_langs) == 1:
                                 sub_name = final_file_name.replace(replacements['ext'], '%s.%s' % (sub_langs[0], replacements['ext']))
                                 rename_files[current_file] = os.path.join(destination, final_folder_name, sub_name)
 
                             rename_files = mergeDicts(rename_files, rename_extras)
 
                         # Filename without cd etc
-                        if file_type is 'movie':
+                        elif file_type is 'movie':
                             rename_extras = self.getRenameExtras(
                                 extra_type = 'movie_extra',
                                 replacements = replacements,
@@ -238,6 +249,7 @@ class Renamer(Plugin):
                             )
                             rename_files = mergeDicts(rename_files, rename_extras)
 
+                            group['filename'] = self.doReplace(file_name, replacements, remove_multiple = True)[:-(len(getExt(final_file_name)) + 1)]
                             group['destination_dir'] = os.path.join(destination, final_folder_name)
 
                         if multiple:
@@ -258,7 +270,7 @@ class Renamer(Plugin):
 
                     # Mark movie "done" onces it found the quality with the finish check
                     try:
-                        if movie.status_id == active_status.get('id'):
+                        if movie.status_id == active_status.get('id') and movie.profile:
                             for profile_type in movie.profile.types:
                                 if profile_type.quality_id == group['meta_data']['quality']['id'] and profile_type.finish:
                                     movie.status_id = done_status.get('id')
@@ -301,7 +313,10 @@ class Renamer(Plugin):
                         elif release.status_id is snatched_status.get('id'):
                             if release.quality.id is group['meta_data']['quality']['id']:
                                 log.debug('Marking release as downloaded')
-                                release.status_id = downloaded_status.get('id')
+                                try:
+                                    release.status_id = downloaded_status.get('id')
+                                except Exception, e:
+                                    log.error('Failed marking release as finished: %s %s', (e, traceback.format_exc()))
                                 db.commit()
 
                 # Remove leftover files
@@ -313,6 +328,7 @@ class Renamer(Plugin):
                     break
 
             # Remove files
+            delete_folders = []
             for src in remove_files:
 
                 if isinstance(src, File):
@@ -324,10 +340,24 @@ class Renamer(Plugin):
 
                 log.info('Removing "%s"', src)
                 try:
-                    os.remove(src)
+                    src = ss(src)
+                    if os.path.isfile(src):
+                        os.remove(src)
+
+                        parent_dir = os.path.normpath(os.path.dirname(src))
+                        if delete_folders.count(parent_dir) == 0 and os.path.isdir(parent_dir) and destination != parent_dir:
+                            delete_folders.append(parent_dir)
+
                 except:
                     log.error('Failed removing %s: %s', (src, traceback.format_exc()))
                     self.tagDir(group, 'failed_remove')
+
+            # Delete leftover folder from older releases
+            for delete_folder in delete_folders:
+                try:
+                    self.deleteEmptyFolder(delete_folder, show_error = False)
+                except Exception, e:
+                    log.error('Failed to delete folder: %s %s', (e, traceback.format_exc()))
 
             # Rename all files marked
             group['renamed_files'] = []
@@ -361,22 +391,22 @@ class Renamer(Plugin):
                 except:
                     log.error('Failed removing %s: %s', (group['parentdir'], traceback.format_exc()))
 
-            # Search for trailers etc
-            fireEventAsync('renamer.after', group)
-
-            # Notify on download
+            # Notify on download, search for trailers etc
             download_message = 'Downloaded %s (%s)' % (movie_title, replacements['quality'])
-            fireEventAsync('movie.downloaded', message = download_message, data = group)
+            try:
+                fireEvent('renamer.after', message = download_message, group = group, in_order = True)
+            except:
+                log.error('Failed firing (some) of the renamer.after events: %s', traceback.format_exc())
 
             # Break if CP wants to shut down
             if self.shuttingDown():
                 break
 
-        #db.close()
         self.renaming_started = False
 
-    def getRenameExtras(self, extra_type = '', replacements = {}, folder_name = '', file_name = '', destination = '', group = {}, current_file = ''):
+    def getRenameExtras(self, extra_type = '', replacements = {}, folder_name = '', file_name = '', destination = '', group = {}, current_file = '', remove_multiple = False):
 
+        replacements = replacements.copy()
         rename_files = {}
 
         def test(s):
@@ -385,8 +415,8 @@ class Renamer(Plugin):
         for extra in set(filter(test, group['files'][extra_type])):
             replacements['ext'] = getExt(extra)
 
-            final_folder_name = self.doReplace(folder_name, replacements)
-            final_file_name = self.doReplace(file_name, replacements)
+            final_folder_name = self.doReplace(folder_name, replacements, remove_multiple = remove_multiple)
+            final_file_name = self.doReplace(file_name, replacements, remove_multiple = remove_multiple)
             rename_files[extra] = os.path.join(destination, final_folder_name, final_file_name)
 
         return rename_files
@@ -418,6 +448,7 @@ class Renamer(Plugin):
                     raise
 
     def moveFile(self, old, dest):
+        dest = ss(dest)
         try:
             shutil.move(old, dest)
 
@@ -426,21 +457,34 @@ class Renamer(Plugin):
             except:
                 log.error('Failed setting permissions for file: %s, %s', (dest, traceback.format_exc(1)))
 
+        except OSError, err:
+            # Copying from a filesystem with octal permission to an NTFS file system causes a permission error.  In this case ignore it.
+            if not hasattr(os, 'chmod') or err.errno != errno.EPERM:
+                raise
+            else:
+                if os.path.exists(dest):
+                    os.unlink(old)
+
         except:
             log.error('Couldn\'t move file "%s" to "%s": %s', (old, dest, traceback.format_exc()))
             raise Exception
 
         return True
 
-    def doReplace(self, string, replacements):
+    def doReplace(self, string, replacements, remove_multiple = False):
         '''
         replace confignames with the real thing
         '''
 
+        replacements = replacements.copy()
+        if remove_multiple:
+            replacements['cd'] = ''
+            replacements['cd_nr'] = ''
+
         replaced = toUnicode(string)
         for x, r in replacements.iteritems():
             if r is not None:
-                replaced = replaced.replace('<%s>' % toUnicode(x), toUnicode(r))
+                replaced = replaced.replace(u'<%s>' % toUnicode(x), toUnicode(r))
             else:
                 #If information is not available, we don't want the tag in the filename
                 replaced = replaced.replace('<' + x + '>', '')
@@ -453,8 +497,10 @@ class Renamer(Plugin):
     def replaceDoubles(self, string):
         return string.replace('  ', ' ').replace(' .', '.')
 
-    def deleteEmptyFolder(self, folder):
+    def deleteEmptyFolder(self, folder, show_error = True):
+        folder = ss(folder)
 
+        loge = log.error if show_error else log.debug
         for root, dirs, files in os.walk(folder):
 
             for dir_name in dirs:
@@ -463,9 +509,91 @@ class Renamer(Plugin):
                     try:
                         os.rmdir(full_path)
                     except:
-                        log.error('Couldn\'t remove empty directory %s: %s', (full_path, traceback.format_exc()))
+                        loge('Couldn\'t remove empty directory %s: %s', (full_path, traceback.format_exc()))
 
         try:
             os.rmdir(folder)
         except:
-            log.error('Couldn\'t remove empty directory %s: %s', (folder, traceback.format_exc()))
+            loge('Couldn\'t remove empty directory %s: %s', (folder, traceback.format_exc()))
+
+    def checkSnatched(self):
+        if self.checking_snatched:
+            log.debug('Already checking snatched')
+
+        self.checking_snatched = True
+
+        snatched_status = fireEvent('status.get', 'snatched', single = True)
+        ignored_status = fireEvent('status.get', 'ignored', single = True)
+        failed_status = fireEvent('status.get', 'failed', single = True)
+
+        done_status = fireEvent('status.get', 'done', single = True)
+
+        db = get_session()
+        rels = db.query(Release).filter_by(status_id = snatched_status.get('id')).all()
+
+        scan_required = False
+
+        if rels:
+            self.checking_snatched = True
+            log.debug('Checking status snatched releases...')
+
+            statuses = fireEvent('download.status', merge = True)
+            if not statuses:
+                log.debug('Download status functionality is not implemented for active downloaders.')
+                scan_required = True
+            else:
+                try:
+                    for rel in rels:
+                        rel_dict = rel.to_dict({'info': {}})
+
+                        # Get current selected title
+                        default_title = getTitle(rel.movie.library)
+
+                        # Check if movie has already completed and is manage tab (legacy db correction)
+                        if rel.movie.status_id == done_status.get('id'):
+                            log.debug('Found a completed movie with a snatched release : %s. Setting release status to ignored...' , default_title)
+                            rel.status_id = ignored_status.get('id')
+                            db.commit()
+                            continue
+
+                        movie_dict = fireEvent('movie.get', rel.movie_id, single = True)
+
+                        # check status
+                        nzbname = self.createNzbName(rel_dict['info'], movie_dict)
+
+                        found = False
+                        for item in statuses:
+                            if item['name'] == nzbname or getImdb(item['name']) == movie_dict['library']['identifier']:
+
+                                timeleft = 'N/A' if item['timeleft'] == -1 else item['timeleft']
+                                log.debug('Found %s: %s, time to go: %s', (item['name'], item['status'].upper(), timeleft))
+
+                                if item['status'] == 'busy':
+                                    pass
+                                elif item['status'] == 'failed':
+                                    fireEvent('download.remove_failed', item, single = True)
+
+                                    if self.conf('next_on_failed'):
+                                        fireEvent('searcher.try_next_release', movie_id = rel.movie_id)
+                                    else:
+                                        rel.status_id = failed_status.get('id')
+                                        db.commit()
+                                elif item['status'] == 'completed':
+                                    log.info('Download of %s completed!', item['name'])
+                                    scan_required = True
+
+                                found = True
+                                break
+
+                        if not found:
+                            log.info('%s not found in downloaders', nzbname)
+
+                except:
+                    log.error('Failed checking for release in downloader: %s', traceback.format_exc())
+
+        if scan_required:
+            fireEvent('renamer.scan')
+
+        self.checking_snatched = False
+
+        return True
